requirements.txt
altair
pandas
streamlit
requests
tqdm
xmltodict
python-dotenv
msal
flask
toml


